Move models into request/response packages
