+++ sso/src/main/java/drewhamilton/skylight/sso/SsoSkylight.kt
@@ -5,8 +5,8 @@ import drewhamilton.skylight.Skylight
 import drewhamilton.skylight.SkylightDay
 import drewhamilton.skylight.sso.network.ApiConstants
 import drewhamilton.skylight.sso.network.InfoClient
-import drewhamilton.skylight.sso.network.Params
-import drewhamilton.skylight.sso.network.SunriseSunsetInfo
+import drewhamilton.skylight.sso.network.response.SunriseSunsetInfo
+import drewhamilton.skylight.sso.network.request.Params
 import java.util.Date
 import javax.inject.Inject
