+++ sso/src/main/java/drewhamilton/skylight/sso/network/SsoApi.kt
@@ -1,5 +1,6 @@
 package drewhamilton.skylight.sso.network
 
+import drewhamilton.skylight.sso.network.response.Response
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
