+++ sso/src/main/java/drewhamilton/skylight/sso/network/response/Response.kt
@@ -1,3 +1,3 @@
-package drewhamilton.skylight.sso.network
+package drewhamilton.skylight.sso.network.response
 
 data class Response(val results: SunriseSunsetInfo, val status: String)
