+++ sso/src/test/java/drewhamilton/skylight/sso/network/InfoClientTest.kt
@@ -2,6 +2,9 @@ package drewhamilton.skylight.sso.network
 
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
+import drewhamilton.skylight.sso.network.request.Params
+import drewhamilton.skylight.sso.network.response.Response
+import drewhamilton.skylight.sso.network.response.SunriseSunsetInfo
 import okhttp3.ResponseBody
 import org.junit.Assert.assertEquals
 import org.junit.Test
@@ -41,7 +44,12 @@ class InfoClientTest {
         mockApi = mock {
             on {
                 getInfo(dummyParams.lat, dummyParams.lng, dummyDateString, 0)
-            } doReturn DummyCall.success(Response(dummySunriseSunsetInfo, "Dummy status"))
+            } doReturn DummyCall.success(
+                Response(
+                    dummySunriseSunsetInfo,
+                    "Dummy status"
+                )
+            )
         }
 
         val infoClient = InfoClient(mockApi, mockDateTimeAdapter)
